Suggested doc edit to explain CSS sizing
I was trying to make my icon size responsive, and it took me quite a bit of work to figure/find out that I could use the CSS font-size attribute. This may be obvious to folks upgrading from FA4 (though maybe not), but it is not well documented. Thanks for your work on this project!